python-faker/requirements.txt
Faker==37.3.0
psycopg2-binary==2.9.10
tqdm==4.67.1
tzdata==2025.2
